file-handle: extract common file io operations
 * ImageFileHandle inherited from FileHandle
 * restore file pointer position after get_file_size
