+++ tests/test_inline.h
@@ -48,132 +48,4 @@ ensure_gpu_buffer_done (SmartPtr<VideoBuffer> buf)
     buf->unmap ();
 }
 
-class FileHandle {
-public:
-    FileHandle ()
-        : _fp (NULL)
-        , _file_name (NULL)
-    {}
-    ~FileHandle ();
-
-    bool is_valid () const {
-        return (_fp ? true : false);
-    }
-    bool end_of_file ();
-    XCamReturn open (const char *name, const char *option);
-    XCamReturn close ();
-    XCamReturn rewind ();
-    XCamReturn get_file_size (size_t &size);
-    XCamReturn read_file (void *buf, const size_t &size);
-    XCamReturn write_file (const void *buf, const size_t &size);
-
-private:
-    XCAM_DEAD_COPY (FileHandle);
-
-private:
-    FILE    *_fp;
-    char    *_file_name;
-};
-
-FileHandle::~FileHandle ()
-{
-    close ();
-}
-
-bool
-FileHandle::end_of_file()
-{
-    if (!is_valid ())
-        return true;
-
-    return feof (_fp);
-}
-
-XCamReturn
-FileHandle::open (const char *name, const char *option)
-{
-    XCAM_ASSERT (name);
-    if (!name)
-        return XCAM_RETURN_ERROR_FILE;
-
-    close ();
-    XCAM_ASSERT (!_file_name && !_fp);
-    _fp = fopen (name, option);
-
-    if (!_fp)
-        return XCAM_RETURN_ERROR_FILE;
-    _file_name = strndup (name, 512);
-
-    return XCAM_RETURN_NO_ERROR;
-}
-
-XCamReturn
-FileHandle::close ()
-{
-    if (_fp) {
-        fclose (_fp);
-        _fp = NULL;
-    }
-
-    if (_file_name) {
-        xcam_free (_file_name);
-        _file_name = NULL;
-    }
-
-    return XCAM_RETURN_NO_ERROR;
-}
-
-XCamReturn
-FileHandle::rewind ()
-{
-    if (!is_valid ())
-        return XCAM_RETURN_ERROR_FILE;
-    return (fseek (_fp, 0L, SEEK_SET) == 0) ? XCAM_RETURN_NO_ERROR : XCAM_RETURN_ERROR_FILE;
-}
-
-XCamReturn
-FileHandle::get_file_size (size_t &size)
-{
-    if (fseek (_fp, 0L, SEEK_END) != 0)
-        goto read_error;
-
-    if ((size = ftell (_fp)) <= 0)
-        goto read_error;
-
-    if (fseek (_fp, 0L, SEEK_SET) != 0)
-        goto read_error;
-
-    return XCAM_RETURN_NO_ERROR;
-
-read_error:
-    XCAM_LOG_ERROR ("get file size failed");
-    return XCAM_RETURN_ERROR_FILE;
-}
-
-XCamReturn
-FileHandle::read_file (void *buf, const size_t &size)
-{
-    XCamReturn ret = XCAM_RETURN_NO_ERROR;
-
-    if (fread (buf, 1, size, _fp) != size) {
-        XCAM_LOG_ERROR ("read file failed, size doesn't match");
-        ret = XCAM_RETURN_ERROR_FILE;
-    }
-
-    return ret;
-}
-
-XCamReturn
-FileHandle::write_file (const void *buf, const size_t &size)
-{
-    XCamReturn ret = XCAM_RETURN_NO_ERROR;
-
-    if (fwrite (buf, 1, size, _fp) != size) {
-        XCAM_LOG_ERROR ("write file failed, size doesn't match");
-        ret = XCAM_RETURN_ERROR_FILE;
-    }
-
-    return ret;
-}
-
 #endif // XCAM_TEST_INLINE_H
