+++ xcore/image_file_handle.h
@@ -22,34 +22,23 @@
 #define XCAM_IMAGE_FILE_HANDLE_H
 
 #include "xcam_utils.h"
+#include "file_handle.h"
 #include "video_buffer.h"
 
 namespace XCam {
 
-class ImageFileHandle {
+class ImageFileHandle
+    : public FileHandle
+{
 public:
-    ImageFileHandle ()
-        : _fp (NULL)
-        , _file_name (NULL)
-    {}
-    ~ImageFileHandle ();
+    ImageFileHandle ();
+    virtual ~ImageFileHandle ();
 
-    bool is_valid () const {
-        return (_fp ? true : false);
-    }
-    bool end_of_file ();
-    XCamReturn open (const char *name, const char *option);
-    XCamReturn close ();
-    XCamReturn rewind ();
     XCamReturn read_buf (const SmartPtr<VideoBuffer> &buf);
     XCamReturn write_buf (const SmartPtr<VideoBuffer> &buf);
 
 private:
     XCAM_DEAD_COPY (ImageFileHandle);
-
-private:
-    FILE   *_fp;
-    char   *_file_name;
 };
 
 }
